requirements.txt
pytest
Appium-Python-Client
python-dotenv